major fix on login controller

+++ app/Http/Controllers/Authentication/LoginController.php
@@ -10,6 +10,8 @@ use App\Models\VerificationCode;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Hash;
+use Illuminate\Validation\ValidationException;
 
 class LoginController extends Controller
 {
@@ -18,22 +20,20 @@ class LoginController extends Controller
         $validate = $request->validated();
         $user = User::where('email', $validate['email'])->first();
 
-        if(Auth::attempt($validate) && $user)
-        {
-            $request->session()->regenerate();
-            
-            return response()->json([
-                'access_token' => $user->createToken('api_token')->plainTextToken,
-                'data' => $user,
-                'token_type' => 'Bearer',
-                "message" => "You have successfully logged in!"
-            ], 200);
+        if (! $user || ! Hash::check($request->password, $user->password)) {
+            throw ValidationException::withMessages([
+                'email' => ['The provided credentials do not match in our records.'],
+            ]);
         }
 
+        $request->session()->regenerate();
+        
         return response()->json([
-            "message" => "The provided credentials do not match in our records."
-        ], 401);
-
+            'access_token' => $user->createToken('api_token')->plainTextToken,
+            'data' => $user,
+            'token_type' => 'Bearer',
+            "message" => "You have successfully logged in!"
+        ], 200);
     }
 
     /**
@@ -46,7 +46,13 @@ class LoginController extends Controller
     {
         $validated = $request->validated();
         $verification_code = VerificationCode::where('otp', $validated['otp'])->first();
-        $user = $verification_code->user; 
+        $user = $verification_code->user?? null; 
+
+        if (!$user) {
+            throw ValidationException::withMessages([
+                'otp' => ['The invalid OTP.'],
+            ]);
+        }
         
         $verification_code->delete();
